# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignEnrollmentReviewerCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "reviewer_id": {
            "format": "uuid",
            "title": "Reviewer Id",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "reviewer_id",
          "reason"
        ],
        "title": "AssignEnrollmentReviewerCommand",
        "type": "object"
      },
...
      "AuditListOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntryOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AuditListOut",
        "type": "object"
      },
      "AuditListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditListOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "AuditListResponse",
        "type": "object"
      },
      "CancelEnrollmentCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "reason"
        ],
        "title": "CancelEnrollmentCommand",
        "type": "object"
      },
...
      "EnrollmentMutationOut": {
        "additionalProperties": false,
        "properties": {
          "idempotency_replay": {
            "default": false,
            "title": "Idempotency Replay",
            "type": "boolean"
          },
          "resource_id": {
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          },
          "reviewer_id": {
            "format": "uuid",
            "title": "Reviewer Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "status",
          "revision",
          "reviewer_id"
        ],
        "title": "EnrollmentMutationOut",
        "type": "object"
      },
      "EnrollmentMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EnrollmentMutationOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "EnrollmentMutationResponse",
        "type": "object"
      },
      "EnrollmentOpsListOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EnrollmentOpsOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "EnrollmentOpsListOut",
        "type": "object"
      },
      "EnrollmentOpsListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EnrollmentOpsListOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "EnrollmentOpsListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuntimeStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RuntimeStatusOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "RuntimeStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Muchen Journey vNext API",
    "version": "0.2.0-local"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ops/audit": {
      "get": {
        "operationId": "list_audit_api_v1_ops_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "resource_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          },
          {
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "in": "query",
            "name": "occurred_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred After"
            }
          },
          {
            "in": "query",
            "name": "occurred_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit"
      }
    },
    "/api/v1/ops/enrollments": {
      "get": {
        "operationId": "list_enrollments_api_v1_ops_enrollments_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentOpsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Enrollments"
      }
    },
    "/api/v1/ops/enrollments/{enrollment_id}/cancel": {
      "post": {
        "operationId": "cancel_enrollment_api_v1_ops_enrollments__enrollment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Enrollment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelEnrollmentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Enrollment"
      }
    },
    "/api/v1/ops/enrollments/{enrollment_id}/reviewer": {
      "put": {
        "operationId": "assign_reviewer_api_v1_ops_enrollments__enrollment_id__reviewer_put",
        "parameters": [
          {
            "in": "path",
            "name": "enrollment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Enrollment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignEnrollmentReviewerCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Reviewer"
      }
    },
...
    "/api/v1/ops/runtime-status": {
      "get": {
        "operationId": "runtime_status_api_v1_ops_runtime_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Status"
      }
    },
...
  }
}